Extract RootTemplate
Extract `RootTemplate` into separate compilation unit to allow granular use in annotation feature template.
